Update CSS
I am adjusting the image so that it is a certain smaller sizee and that it is at the center of the page. I have had to add the display:block; in order for the image to be treated like a block level element so that I can manipulate it to the position and size I want it to be at.