Update CSS
I applied animation-name: pulse to the .active-day elements.
I made the duration of the animation 2s and ensure it plays infinite times.
I used @keyframes for the pulse animation to make the text size change from 100% and 120% and back again, and change the color to #00CC00 at the 50% mark of the animation.

Although I have succeeded in making the text itslf pulse, they are not in a column. I might come back to fix this.